+++ dpws/src/main/java/org/somda/sdc/dpws/CommunicationLogContext.java
@@ -1,0 +1,21 @@
+package org.somda.sdc.dpws;
+
+/**
+ * Additional information for {@linkplain CommunicationLog} implementations.
+ */
+public class CommunicationLogContext {
+    private String eprAddress;
+
+    public CommunicationLogContext(String eprAddress) {
+        this.eprAddress = eprAddress;
+    }
+
+    /**
+     * Gets the EPR address (identification) of a device.
+     *
+     * @return the EPR address.
+     */
+    public String getEprAddress() {
+        return eprAddress;
+    }
+}