+++ dpws/src/main/java/org/somda/sdc/dpws/CommunicationLogDummyImpl.java
@@ -1,7 +1,10 @@
 package org.somda.sdc.dpws;
 
+import com.google.inject.assistedinject.Assisted;
+import com.google.inject.assistedinject.AssistedInject;
 import org.somda.sdc.dpws.soap.CommunicationContext;
 
+import javax.annotation.Nullable;
 import java.io.InputStream;
 import java.io.OutputStream;
 
@@ -9,22 +12,37 @@ import java.io.OutputStream;
  * Implementation of {@link CommunicationLog} that does not output anything.
  */
 public class CommunicationLogDummyImpl implements CommunicationLog {
-    
+    @AssistedInject
+    CommunicationLogDummyImpl() {
+    }
+
+    @AssistedInject
+    CommunicationLogDummyImpl(@Assisted @Nullable CommunicationLogContext communicationLogContext) {
+    }
+
     @Override
-    public OutputStream logMessage(Direction direction, TransportType transportType, MessageType messageType,
-                                   CommunicationContext communicationContext, OutputStream message) {
+    public OutputStream logMessage(Direction direction,
+                                   TransportType transportType,
+                                   MessageType messageType,
+                                   CommunicationContext communicationContext,
+                                   OutputStream message) {
         return message;
     }
 
     @Override
-    public OutputStream logMessage(Direction direction, TransportType transportType, MessageType messageType,
+    public OutputStream logMessage(Direction direction,
+                                   TransportType transportType,
+                                   MessageType messageType,
                                    CommunicationContext communicationContext) {
         return OutputStream.nullOutputStream();
     }
 
     @Override
-    public InputStream logMessage(Direction direction, TransportType transportType, MessageType messageType,
-                                  CommunicationContext communicationContext, InputStream message) {
+    public InputStream logMessage(Direction direction,
+                                  TransportType transportType,
+                                  MessageType messageType,
+                                  CommunicationContext communicationContext,
+                                  InputStream message) {
         return message;
     }
 }